refactor: change a word Account -> User
Account 에서 User로 단어변경
