+++ app/models/__init__.py
@@ -1,4 +1,4 @@
 # 모든 모델을 여기에서 import해야 Alembic이 인식할 수 있습니다
-from app.models.account import Account
+from app.models.user import User
 
-__all__ = ["Account"]
+__all__ = ["User"]
